+++ Assets/Scripts/MainMenuManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8191e62c6d032d24888c4fe40ba0874c
+guid: 6de95521a80e3474f98377ef107b042c
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
