+++ Assets/Scripts/MenuManager.cs
@@ -39,6 +39,6 @@ public class MenuManager : MonoBehaviour
     public void QuitGame()
     {
         Application.Quit();
-        // UnityEditor.EditorApplication.isPlaying = false; // Uncomment if testing in the Unity Editor
+        UnityEditor.EditorApplication.isPlaying = false; // Uncomment if testing in the Unity Editor
     }
 }
