--- Assets/Scripts/StartButton.cs
@@ -1,22 +1,0 @@
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.SceneManagement;
-
-public class StartButton : MonoBehaviour
-{
-    // Start is called before the first frame update
-    void Start()
-    {
-
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-
-    }
-    public void GameStart()
-    {
-        SceneManager.LoadScene("SampleScene");
-    }
-}
