--- Assets/Scripts/StartButton.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ef9690ecff24cb546813590c8ed7d21f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
