--- Assets/Scripts/TutorialButton.cs
@@ -1,22 +1,0 @@
-using UnityEngine;
-using UnityEngine.SceneManagement;
-
-public class TutorialButton : MonoBehaviour
-{
-    // Start is called before the first frame update
-    void Start()
-    {
-
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-
-    }
-    public void Tutorial()
-    {
-        SceneManager.LoadScene("SplashScreen");
-    }
-}
-
